WIP: make infrastructure factor and surface type, comfort factor, gradients bindable through BicycleParams interface.
