+++ contribs/bicycle/src/main/java/org/matsim/contrib/bicycle/AdditionalBicycleLinkScoreDefaultImpl.java
@@ -7,6 +7,8 @@ import org.matsim.core.config.ConfigUtils;
 
 public final class AdditionalBicycleLinkScoreDefaultImpl implements AdditionalBicycleLinkScore {
 
+	@Inject
+	private BicycleParams bicycleParams;
 	private final double marginalUtilityOfInfrastructure_m;
 	private final double userDefinedNetworkAttributeDefaultValue;
 	private final double marginalUtilityOfComfort_m;
@@ -30,13 +32,13 @@ public final class AdditionalBicycleLinkScoreDefaultImpl implements AdditionalBi
 
 		double distance = link.getLength();
 
-		double comfortFactor = BicycleUtils.getComfortFactor(surface );
+		double comfortFactor = bicycleParams.getComfortFactor(surface );
 		double comfortScore = marginalUtilityOfComfort_m * (1. - comfortFactor) * distance;
 
-		double infrastructureFactor = BicycleUtils.getInfrastructureFactor(type, cyclewaytype );
+		double infrastructureFactor = bicycleParams.getInfrastructureFactor(type, cyclewaytype );
 		double infrastructureScore = marginalUtilityOfInfrastructure_m * (1. - infrastructureFactor) * distance;
 
-		double gradient = BicycleUtils.getGradient( link );
+		double gradient = bicycleParams.getGradient( link );
 		double gradientScore = marginalUtilityOfGradient_m_100m * gradient * distance;
 
 		String userDefinedNetworkAttributeString;
