+++ contribs/bicycle/src/main/java/org/matsim/contrib/bicycle/BicycleModule.java
@@ -74,6 +74,11 @@ public final class BicycleModule extends AbstractModule {
 		this.bind( AdditionalBicycleLinkScore.class ).to( AdditionalBicycleLinkScoreDefaultImpl.class );
 		// (this computes the value of the per-link scoring event.  yyyy Very unfortunately, it is a re-implementation of the BicycleTravelDisutility (mentioned above).)
 
+
+//		here, different values for surface types and infrastructure factors can be defined.
+//		by default, the default values in BicycleParamsDefaultImpl are bound
+		this.bind(BicycleParams.class).to(BicycleParamsDefaultImpl.class);
+
 		this.installOverridingQSimModule( new AbstractQSimModule(){
 			@Override protected void configureQSim(){
 				this.addLinkSpeedCalculator().to( BicycleLinkSpeedCalculator.class );
