+++ contribs/bicycle/src/main/java/org/matsim/contrib/bicycle/BicycleParams.java
@@ -1,0 +1,11 @@
+package org.matsim.contrib.bicycle;
+
+import org.matsim.api.core.v01.network.Link;
+
+public interface BicycleParams {
+	public double getComfortFactor(String surface);
+
+	double getInfrastructureFactor(String type, String cyclewaytype);
+
+	double getGradient(Link link);
+}
