+++ contribs/bicycle/src/test/java/org/matsim/contrib/bicycle/BicycleDefaultImplTest.java
@@ -1,0 +1,10 @@
+package org.matsim.contrib.bicycle;
+
+// Test everything in BicycleParamsDefaultImpl
+public class BicycleDefaultImplTest {
+
+
+
+
+
+}
